Refactor Texture validation to it's own class
