+++ src/main/java/com/dumbdogdiner/stickyapi/common/util/textures/TextureHelper.java
@@ -10,22 +10,14 @@ import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import lombok.experimental.UtilityClass;
-import okhttp3.HttpUrl;
 import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.Response;
-import org.apache.commons.lang.NullArgumentException;
-import org.apache.commons.validator.routines.UrlValidator;
 import org.jetbrains.annotations.NotNull;
 import org.yaml.snakeyaml.Yaml;
 
-import javax.imageio.ImageIO;
-import java.io.*;
-import java.net.MalformedURLException;
+import java.io.IOException;
+import java.io.InputStream;
 import java.net.URL;
-import java.nio.charset.StandardCharsets;
 import java.util.*;
-import java.util.logging.Level;
 
 /**
  * Contains various helper methods related to Textures, Texture Strings, and texture validation, as well as providing
@@ -34,11 +26,8 @@ import java.util.logging.Level;
 //TODO write javadoc
 @UtilityClass
 public class TextureHelper {
-    /**
-     * Byte array containing the file signature of a PNG image
-     */
-    private static final byte[] PNG_SIGNATURE = {(byte) 0x89, (byte) 0x50, (byte) 0x4e, (byte) 0x47, (byte) 0x0d, (byte) 0x0a, (byte) 0x1a, (byte) 0x0a};
-    private static final OkHttpClient httpClient = new OkHttpClient();
+    // Package-local visibility
+    static final OkHttpClient httpClient = new OkHttpClient();
 
     private static Map<String, Map<String, String>> TextureMap;
     private static final Yaml YAML = new Yaml();
@@ -175,64 +164,7 @@ public class TextureHelper {
     }
 
 
-    public static boolean isValidTextureUrl(String url) {
-        //TODO implement
-        return true;
-    }
 
-    /**
-     * Attempts to (as exhaustively as possible) validate a given String of a Base-64 encoded texture
-     *
-     * @param texture The string that represents the texture
-     * @return if the texture is valid or not
-     */
-    public static boolean isValidTexture(String texture) {
-        try {
-            String decodedTexture = new String(Base64.getDecoder().decode(texture), StandardCharsets.UTF_8);
-            String textureURL = JsonParser.parseString(decodedTexture)
-                    .getAsJsonObject().get("textures")
-                    .getAsJsonObject().get("SKIN")
-                    .getAsJsonObject().get("url")
-                    .getAsString();
-            if (!UrlValidator.getInstance().isValid(textureURL)) {
-                throw new MalformedURLException("The texture URL " + textureURL + " is not a valid URL");
-            }
-            String host = HttpUrl.get(textureURL).host().toLowerCase();
-            if (!(host.endsWith("mojang.com") || host.endsWith("minecraft.net"))) {
-                throw new MalformedURLException("The texture URL " + textureURL + " specified a host other than minecraft.net or mojang.com");
-            }
-            Response resp = httpClient.newCall(new Request.Builder().url(textureURL).build()).execute();
-            if (resp.code() != 200) {
-                resp.close();
-                throw new Exception("Non 200/OK Response of " + resp.code() + " received from " + textureURL);
-            }
-            if (!"image/png".equalsIgnoreCase(resp.header("Content-Type"))) {
-                resp.close();
-                throw new Exception("Unexpected format of " + resp.header("Content-Type") + " was received from " + textureURL);
-            }
-            try {
-                byte[] response = resp.body().bytes();
-                if (ImageIO.read(new ByteArrayInputStream(response)) != null) {
-                    for (int i = 0; i < PNG_SIGNATURE.length; i++) {
-                        if (PNG_SIGNATURE[i] != response[i]) {
-                            int j = i + 1;
-                            throw new Exception("Byte " + j + " of the response (" + response[i] + ") did not match byte " + j + " of the PNG signature (" + PNG_SIGNATURE[i] + ")");
-                        }
-                    }
-                    return true;
-                } else {
-                    throw new NullArgumentException("The image retrieved from " + textureURL + " was decoded to null and" /* must not be null */);
-                }
-            } catch (IOException | NullArgumentException | NullPointerException e) {
-                throw new Exception("The content retrieved from " + textureURL + " was not a recognized image, was null, or was decoded to null", e);
-            }
-        } catch (Exception e) {
-            StringWriter sw = new StringWriter();
-            e.printStackTrace(new PrintWriter(sw));
-            StickyAPI.getLogger().log(Level.INFO, sw.toString());
-            return false;
-        }
-    }
 
     public static String toQualifiedName(String filter, String head) {
         return (filter + '.' + head).toUpperCase();
