+++ src/test/java/com/dumbdogdiner/stickyapi/common/util/textures/TextureValidatorTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class TextureHelperTest {
+class TextureValidatorTest {
     @BeforeAll
     static void setUp() {
         TestsCommon.disableHandlers();
@@ -30,7 +30,7 @@ class TextureHelperTest {
 
     @Test
     void validateTexture() {
-        assertTrue(TextureHelper.isValidTexture("eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvZDViM2Y4Y2E0YjNhNTU1Y2NiM2QxOTQ0NDk4MDhiNGM5ZDc4MzMyNzE5NzgwMGQ0ZDY1OTc0Y2M2ODVhZjJlYSJ9fX0="));
+        assertTrue(TextureValidator.isValidTexture("eyJ0ZXh0dXJlcyI6eyJTS0lOIjp7InVybCI6Imh0dHA6Ly90ZXh0dXJlcy5taW5lY3JhZnQubmV0L3RleHR1cmUvZDViM2Y4Y2E0YjNhNTU1Y2NiM2QxOTQ0NDk4MDhiNGM5ZDc4MzMyNzE5NzgwMGQ0ZDY1OTc0Y2M2ODVhZjJlYSJ9fX0="));
     }
 
     @Test
@@ -42,7 +42,7 @@ class TextureHelperTest {
         };
         for (String invalidTest : tests) {
             System.out.println(invalidTest);
-            assertFalse(TextureHelper.isValidTexture(invalidTest));
+            assertFalse(TextureValidator.isValidTexture(invalidTest));
         }
     }
 
@@ -122,7 +122,7 @@ class TextureHelperTest {
     private void testMultiFail(String[] tests) {
         for (String invalidTest : tests) {
             System.out.println("Now Testing invalid texture " + invalidTest);
-            assertFalse(TextureHelper.isValidTexture(invalidTest));
+            assertFalse(TextureValidator.isValidTexture(invalidTest));
         }
     }
 
@@ -143,7 +143,7 @@ class TextureHelperTest {
     void validateAllTextures() {
         for(String qn : TextureHelper.getQualifiedNames()){
             System.out.println("Testing texture for " + qn + " (" + TextureHelper.getTexture(qn) + ")");
-            assertTrue(TextureHelper.isValidTexture(TextureHelper.getTexture(qn)));
+            assertTrue(TextureValidator.isValidTexture(TextureHelper.getTexture(qn)));
         }
     }
